// src/infrastructure/kot/types.ts
// KOT sort index values used as data-ht-sort-index attribute selectors
export type KotSortIndex =
  | "ALL_WORK_MINUTE"
  | "FIXED_WORK_MINUTE"
  | "OVERTIME_WORK_MINUTE"
  | "REST_MINUTE"
  | "START_TIMERECORD"
  | "END_TIMERECORD"
  | "REST_START_TIMERECORD"
  | "REST_END_TIMERECORD"
  | "WORK_DAY"
  | "WORK_DAY_TYPE"
  | "SCHEDULE";